requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
faker
google-genai
google-cloud-bigquery
pydantic
tenacity
python-dotenv